+++ Program.cs
@@ -1,5 +1,6 @@
 using FMSD_BE.Data;
 using FMSD_BE.Services.DashboardService;
+using FMSD_BE.Services.FilterService;
 using FMSD_BE.Services.ReportService.AlarmService;
 using FMSD_BE.Services.ReportService.CalibrationDetailService;
 using FMSD_BE.Services.ReportService.CalibrationService;
@@ -19,6 +20,7 @@ builder.Services.AddControllers();
 builder.Services.AddSwaggerGen();
 builder.Services.AddDbContext<CentralizedFmsCloneContext>(options => options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
 builder.Services.AddScoped<IDashboardService, DashboardService>();
+builder.Services.AddScoped<IFilterService, FilterService>();
 builder.Services.AddScoped<IAlarmService, AlarmService>();
 builder.Services.AddScoped<ITankService, TankService>();
 builder.Services.AddScoped<ISharedService, SharedService>();
@@ -30,10 +32,10 @@ builder.Services.AddScoped<ICalibrationDetailService, CalibrationDetailService>(
 
 builder.Services.AddCors(o => o.AddPolicy("CorsPolicy", builder =>
 {
-    builder.AllowAnyMethod()
-            .AllowAnyHeader()
-            .SetIsOriginAllowed(origin => true) // allow any origin you can change here to allow localhost:4200
-            .AllowCredentials();
+	builder.AllowAnyMethod()
+			.AllowAnyHeader()
+			.SetIsOriginAllowed(origin => true) // allow any origin you can change here to allow localhost:4200
+			.AllowCredentials();
 }));
